type safety

+++ Core/MemoryFile.h
@@ -100,7 +100,7 @@ extern bool mkPath(const MMKVPath_t &path);
 extern bool isFileExist(const MMKVPath_t &nsFilePath);
 extern MMBuffer *readWholeFile(const MMKVPath_t &path);
 extern bool zeroFillFile(MMKVFileHandle_t fd, size_t startPos, size_t size);
-extern int getPageSize();
+extern size_t getPageSize();
 } // namespace mmkv
 
 #endif //MMKV_MAMERYFILE_H
